requirements.txt
telethon
requests
python-dotenv
transformers
torch
sqlite3-api
python-telegram-bot>=20.0 